+++ src/Octokit.Webhooks/Models/PullRequestEvent/PullRequest.cs
@@ -14,7 +14,7 @@
         public string Url { get; init; } = null!;
 
         [JsonPropertyName("id")]
-        public int Id { get; init; }
+        public long Id { get; init; }
 
         [JsonPropertyName("node_id")]
         public string NodeId { get; init; } = null!;
@@ -32,7 +32,7 @@
         public string IssueUrl { get; init; } = null!;
 
         [JsonPropertyName("number")]
-        public int Number { get; init; }
+        public long Number { get; init; }
 
         [JsonPropertyName("state")]
         public PullRequestState State { get; init; }
@@ -131,24 +131,24 @@
         public User? MergedBy { get; init; }
 
         [JsonPropertyName("comments")]
-        public int Comments { get; init; }
+        public long Comments { get; init; }
 
         [JsonPropertyName("review_comments")]
-        public int ReviewComments { get; init; }
+        public long ReviewComments { get; init; }
 
         [JsonPropertyName("maintainer_can_modify")]
         public bool MaintainerCanModify { get; init; }
 
         [JsonPropertyName("commits")]
-        public int Commits { get; init; }
+        public long Commits { get; init; }
 
         [JsonPropertyName("additions")]
-        public int Additions { get; init; }
+        public long Additions { get; init; }
 
         [JsonPropertyName("deletions")]
-        public int Deletions { get; init; }
+        public long Deletions { get; init; }
 
         [JsonPropertyName("changed_files")]
-        public int ChangedFiles { get; init; }
+        public long ChangedFiles { get; init; }
     }
 }
