+++ src/Octokit.Webhooks/Models/SecretScanningAlertEvent/Alert.cs
@@ -7,7 +7,7 @@
     public sealed record Alert
     {
         [JsonPropertyName("number")]
-        public int Number { get; init; }
+        public long Number { get; init; }
 
         [JsonPropertyName("secret_type")]
         public string SecretType { get; init; } = null!;
